[AI] Set up initial server app with registration
Claude 3.7 prompt:

---

I need to add a simple web server. It should support the following
endpoint, which will be called by a client device:

    curl "http://localhost:8000/api/setup/" \
        -H 'ID: <redacted>' \
        -H 'Access-Token: <redacted>' \
        -H 'Accept: application/json' \
        -H 'Content-Type: application/json'

The ID is the MAC address of the device. The Access-Token is a hardcoded
secret that can be put as a const into the source code. The server
should verify that the access token matches the server-side value.

When the setup endpoint is called, a new device should be registered and
stored in an SQLite database.